implemented more methods and enhanced security

+++ src/main/java/com/arthuurdp/e_commerce/entities/dtos/cart_item/CartItemResponse.java
@@ -1,4 +1,6 @@
 package com.arthuurdp.e_commerce.entities.dtos.cart_item;
 
-public record CartItemResponse(Long id, String name, Double price, Integer quantity, Double subtotal) {
+import java.math.BigDecimal;
+
+public record CartItemResponse(Long id, String name, BigDecimal price, Integer quantity, BigDecimal subtotal) {
 }
